fix(seeder): database seeder and remove permission tables

+++ database/migrations/2026_05_08_125504_create_pendapatan.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('pendapatan', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('transaksi_id');
+            $table->unsignedBigInteger('transaksi_id');
             $table->date('tanggal');
             $table->decimal('keuntungan', 12, 2)->default(0);
             $table->timestamps();
